requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiohttp==3.9.3
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   langchain-core
    #   openai
    #   sse-starlette
    #   starlette
asttokens==2.4.1
    # via stack-data
attrs==23.2.0
    # via aiohttp
authlib==1.3.0
    # via weaviate-client
beautifulsoup4==4.12.3
    # via -r requirements.in
black==23.7.0
    # via -r requirements.in
build==1.0.3
    # via pip-tools
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   pip-tools
    #   typer
    #   uvicorn
colorama==0.4.6
    # via typer
cryptography==42.0.4
    # via authlib
dataclasses-json==0.6.4
    # via
    #   langchain
    #   langchain-community
decorator==5.1.1
    # via ipython
distro==1.9.0
    # via openai
executing==2.0.1
    # via stack-data
fastapi==0.109.2
    # via
    #   langserve
    #   sse-starlette
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.11
    # via gitpython
gitpython==3.1.42
    # via langchain-cli
greenlet==3.0.3
    # via sqlalchemy
grpcio==1.62.0
    # via
    #   grpcio-health-checking
    #   grpcio-tools
    #   weaviate-client
grpcio-health-checking==1.62.0
    # via weaviate-client
grpcio-tools==1.62.0
    # via weaviate-client
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.4
    # via httpx
httpx==0.26.0
    # via
    #   langserve
    #   openai
    #   weaviate-client
httpx-sse==0.4.0
    # via langserve
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
ipython==8.22.1
    # via -r requirements.in
jedi==0.19.1
    # via ipython
jsonpatch==1.33
    # via
    #   langchain
    #   langchain-core
jsonpointer==2.4
    # via jsonpatch
langchain==0.1.8
    # via
    #   -r requirements.in
    #   langserve
langchain-cli==0.0.21
    # via -r requirements.in
langchain-community==0.0.21
    # via langchain
langchain-core==0.1.25
    # via
    #   langchain
    #   langchain-community
    #   langchain-openai
langchain-openai==0.0.6
    # via -r requirements.in
langchainhub==0.1.14
    # via -r requirements.in
langserve[all]==0.0.42
    # via
    #   -r requirements.in
    #   langchain-cli
langsmith==0.1.5
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markdown-it-py==3.0.0
    # via rich
marshmallow==3.20.2
    # via dataclasses-json
matplotlib-inline==0.1.6
    # via ipython
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via
    #   black
    #   typing-inspect
numpy==1.26.4
    # via
    #   langchain
    #   langchain-community
    #   langchain-openai
openai==1.12.0
    # via langchain-openai
orjson==3.9.14
    # via langserve
packaging==23.2
    # via
    #   black
    #   build
    #   langchain-core
    #   marshmallow
parso==0.8.3
    # via jedi
pathspec==0.12.1
    # via black
pexpect==4.9.0
    # via ipython
pip-tools==7.4.0
    # via -r requirements.in
platformdirs==4.2.0
    # via black
prompt-toolkit==3.0.43
    # via ipython
protobuf==4.25.3
    # via
    #   grpcio-health-checking
    #   grpcio-tools
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pydantic==2.6.1
    # via
    #   fastapi
    #   langchain
    #   langchain-core
    #   langserve
    #   langsmith
    #   openai
    #   weaviate-client
pydantic-core==2.16.2
    # via pydantic
pygments==2.17.2
    # via
    #   ipython
    #   rich
pypdf==4.0.2
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via
    #   build
    #   pip-tools
pyyaml==6.0.1
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
regex==2023.12.25
    # via tiktoken
requests==2.31.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
    #   langchainhub
    #   langsmith
    #   tiktoken
    #   weaviate-client
rich==13.7.0
    # via typer
shellingham==1.5.4
    # via typer
six==1.16.0
    # via asttokens
smmap==5.0.1
    # via gitdb
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
    #   openai
soupsieve==2.5
    # via beautifulsoup4
sqlalchemy==2.0.27
    # via
    #   langchain
    #   langchain-community
sse-starlette==1.8.2
    # via langserve
stack-data==0.6.3
    # via ipython
starlette==0.36.3
    # via
    #   fastapi
    #   sse-starlette
tenacity==8.2.3
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tiktoken==0.6.0
    # via langchain-openai
tomlkit==0.12.3
    # via langchain-cli
tqdm==4.66.2
    # via openai
traitlets==5.14.1
    # via
    #   ipython
    #   matplotlib-inline
typer[all]==0.9.0
    # via langchain-cli
types-requests==2.31.0.20240218
    # via langchainhub
typing-extensions==4.9.0
    # via
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.2.1
    # via
    #   requests
    #   types-requests
uvicorn==0.23.2
    # via
    #   langchain-cli
    #   sse-starlette
validators==0.22.0
    # via weaviate-client
wcwidth==0.2.13
    # via prompt-toolkit
weaviate-client==4.4.4
    # via -r requirements.in
wheel==0.42.0
    # via pip-tools
yarl==1.9.4
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
